Delete README.md
# Getting and Cleaning Data Project

This project used the Human Activity Recognition using Smartphones Dataset to demonstrate how to use a variety of tools to obtain and clean data.

The data was transformed to meet project requirements such as mergings the training and the test sets to create one data set, extracting only the measurements on the mean and standard deviation for each measurement, using descriptive activity names to name the activities in the data set, appropriately labeling the data set with descriptive variable names and finally creating a tidy data set with the average of each variable for each activity and each subject. 

The CodeBook.md file explains the transformations performed and the resulting data and variables.